fix completions v2
