+++ routes.gen.ts
@@ -6,28 +6,20 @@ import * as $0 from "./pages/index.tsx";
 import * as $1 from "./pages/std.tsx";
 import * as $2 from "./pages/_render.ts";
 import * as $3 from "./pages/install.ts";
-import * as $4 from "./pages/completions/v2/details_x_pkg_ver_path.ts";
-import * as $5 from "./pages/completions/v2/details_x_pkg_ver.ts";
-import * as $6 from "./pages/completions/v2/x_pkg_ver_path.ts";
-import * as $7 from "./pages/completions/v2/details_std_ver.ts";
-import * as $8 from "./pages/completions/v2/details_x_pkg.ts";
-import * as $9 from "./pages/completions/v2/details_std_ver_path.ts";
-import * as $10 from "./pages/completions/v2/x_pkg_ver.ts";
-import * as $11 from "./pages/completions/v2/x_pkg.ts";
-import * as $12 from "./pages/completions/v1.ts";
-import * as $13 from "./pages/completions/config.ts";
-import * as $14 from "./pages/legacy_redirects/posts.ts";
-import * as $15 from "./pages/legacy_redirects/v1.ts";
-import * as $16 from "./pages/benchmarks.tsx";
-import * as $17 from "./pages/add_module.tsx";
-import * as $18 from "./pages/manual.tsx";
-import * as $19 from "./pages/translations.tsx";
-import * as $20 from "./pages/status.tsx";
-import * as $21 from "./pages/_404.tsx";
-import * as $22 from "./pages/_app.tsx";
-import * as $23 from "./pages/x/index.tsx";
-import * as $24 from "./pages/x/module.tsx";
-import * as $25 from "./pages/artwork.tsx";
+import * as $4 from "./pages/completions/v1.ts";
+import * as $5 from "./pages/completions/config.ts";
+import * as $6 from "./pages/legacy_redirects/posts.ts";
+import * as $7 from "./pages/legacy_redirects/v1.ts";
+import * as $8 from "./pages/benchmarks.tsx";
+import * as $9 from "./pages/add_module.tsx";
+import * as $10 from "./pages/manual.tsx";
+import * as $11 from "./pages/translations.tsx";
+import * as $12 from "./pages/status.tsx";
+import * as $13 from "./pages/_404.tsx";
+import * as $14 from "./pages/_app.tsx";
+import * as $15 from "./pages/x/index.tsx";
+import * as $16 from "./pages/x/module.tsx";
+import * as $17 from "./pages/artwork.tsx";
 
 const routes = {
   pages: {
@@ -35,28 +27,20 @@ const routes = {
     "./pages/std.tsx": $1,
     "./pages/_render.ts": $2,
     "./pages/install.ts": $3,
-    "./pages/completions/v2/details_x_pkg_ver_path.ts": $4,
-    "./pages/completions/v2/details_x_pkg_ver.ts": $5,
-    "./pages/completions/v2/x_pkg_ver_path.ts": $6,
-    "./pages/completions/v2/details_std_ver.ts": $7,
-    "./pages/completions/v2/details_x_pkg.ts": $8,
-    "./pages/completions/v2/details_std_ver_path.ts": $9,
-    "./pages/completions/v2/x_pkg_ver.ts": $10,
-    "./pages/completions/v2/x_pkg.ts": $11,
-    "./pages/completions/v1.ts": $12,
-    "./pages/completions/config.ts": $13,
-    "./pages/legacy_redirects/posts.ts": $14,
-    "./pages/legacy_redirects/v1.ts": $15,
-    "./pages/benchmarks.tsx": $16,
-    "./pages/add_module.tsx": $17,
-    "./pages/manual.tsx": $18,
-    "./pages/translations.tsx": $19,
-    "./pages/status.tsx": $20,
-    "./pages/_404.tsx": $21,
-    "./pages/_app.tsx": $22,
-    "./pages/x/index.tsx": $23,
-    "./pages/x/module.tsx": $24,
-    "./pages/artwork.tsx": $25,
+    "./pages/completions/v1.ts": $4,
+    "./pages/completions/config.ts": $5,
+    "./pages/legacy_redirects/posts.ts": $6,
+    "./pages/legacy_redirects/v1.ts": $7,
+    "./pages/benchmarks.tsx": $8,
+    "./pages/add_module.tsx": $9,
+    "./pages/manual.tsx": $10,
+    "./pages/translations.tsx": $11,
+    "./pages/status.tsx": $12,
+    "./pages/_404.tsx": $13,
+    "./pages/_app.tsx": $14,
+    "./pages/x/index.tsx": $15,
+    "./pages/x/module.tsx": $16,
+    "./pages/artwork.tsx": $17,
   },
   baseUrl: import.meta.url,
 };
