+++ server_deps.ts
@@ -7,3 +7,4 @@ export { prettyBytes } from "https://deno.land/x/pretty_bytes@v1.0.5/mod.ts";
 export { serve } from "https://deno.land/std@0.126.0/http/server.ts";
 export type { ConnInfo } from "https://deno.land/std@0.126.0/http/server.ts";
 export { createReporter } from "https://deno.land/x/g_a@0.1.2/mod.ts";
+export { router } from "https://crux.land/router@0.0.5";
